Create and update a choice
Added functionality to create and update a choice. Choices wont actually be able to be updated, but as an example is provided incase. Updated choice table name, removed confidence column, added cascading to on delete as it cascades to delete the choices record if the reference object is deleted. Updated api end points