Populate Images endpoint, fixes
Added new populate images endpoint to insert image records in a way to keep 20 out of 200 images intended just for the ML model to train on, while the rest of the 180 images are for users to label. Fixed choice serializer to conform to its model attributes by including user_training_record. Url api path fix for the given username to match the methods parameter